+++ lending/CodatLending/CodatLending.cs
@@ -17,6 +17,8 @@ namespace Codat.Lending
     using System.Threading.Tasks;
     using System;
 
+
+
     /// <summary>
     /// Lending API: Our Lending API helps you make smarter credit decisions on small businesses by enabling you to pull your customers&apos; latest data from accounting, banking, and commerce platforms they are already using. It also includes features to help providers verify the accuracy of data and process it more efficiently.<br/>
     /// 
@@ -103,6 +105,22 @@ namespace Codat.Lending
     
     public class SDKConfig
     {
+        public static string[] ServerList = new string[]
+        {
+            "https://api.codat.io",
+        };
+        /// Contains the list of servers available to the SDK
+        public string serverUrl = "";
+        public int serverIndex = 0;
+
+        public string GetTemplatedServerDetails()
+        {
+            if (!String.IsNullOrEmpty(this.serverUrl))
+            {
+                return Utilities.TemplateUrl(Utilities.RemoveSuffix(this.serverUrl, "/"), new Dictionary<string, string>());
+            }
+            return Utilities.TemplateUrl(SDKConfig.ServerList[this.serverIndex], new Dictionary<string, string>());
+        }
     }
 
     /// <summary>
@@ -140,16 +158,12 @@ namespace Codat.Lending
     public class CodatLending: ICodatLending
     {
         public SDKConfig Config { get; private set; }
-        public static List<string> ServerList = new List<string>()
-        {
-            "https://api.codat.io",
-        };
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "5.0.0";
-        private const string _sdkGenVersion = "2.188.1";
+        private const string _sdkVersion = "5.0.1";
+        private const string _sdkGenVersion = "2.194.1";
         private const string _openapiDocVersion = "3.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 5.0.0 2.188.1 3.0.0 Codat.Lending";
+        private const string _userAgent = "speakeasy-sdk/csharp 5.0.1 2.194.1 3.0.0 Codat.Lending";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -170,9 +184,14 @@ namespace Codat.Lending
         public IDataIntegrity DataIntegrity { get; private set; }
         public IExcelReports ExcelReports { get; private set; }
 
-        public CodatLending(Security? security = null, string? serverUrl = null, ISpeakeasyHttpClient? client = null)
+        public CodatLending(Security? security = null, int? serverIndex = null, string? serverUrl = null, Dictionary<string, string>? urlParams = null, ISpeakeasyHttpClient? client = null)
         {
-            _serverUrl = serverUrl ?? CodatLending.ServerList[0];
+            if (serverUrl != null) {
+                if (urlParams != null) {
+                    serverUrl = Utilities.TemplateUrl(serverUrl, urlParams);
+                }
+                _serverUrl = serverUrl;
+            }
 
             _defaultClient = new SpeakeasyHttpClient(client);
             _securityClient = _defaultClient;
@@ -184,6 +203,7 @@ namespace Codat.Lending
             
             Config = new SDKConfig()
             {
+                serverUrl = _serverUrl
             };
 
             Companies = new Companies(_defaultClient, _securityClient, _serverUrl, Config);