+++ lending/RELEASES.md
@@ -173,3 +173,13 @@ Based on:
 - [csharp v5.0.0] lending
 ### Releases
 - [NuGet v5.0.0] https://www.nuget.org/packages/Codat.Lending/5.0.0 - lending
+
+## 2023-11-20 01:28:16
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v5.0.1] lending
+### Releases
+- [NuGet v5.0.1] https://www.nuget.org/packages/Codat.Lending/5.0.1 - lending